Option to cache samples.

+++ Orange/data/sql/table.py
@@ -664,30 +664,37 @@ class SqlTable(table.Table):
     def quote_string(self, value):
         return "'%s'" % value
 
-    def sample_percentage(self, percentage):
+    def sample_percentage(self, percentage, no_cache=False):
         return self._sample('blocksample_percent', percentage)
 
-    def sample_time(self, time_in_seconds):
+    def sample_time(self, time_in_seconds, no_cache=False):
         return self._sample('blocksample_time', int(time_in_seconds * 1000))
 
-    def _sample(self, method, parameter):
+    def _sample(self, method, parameter, no_cache=False):
         if "," in self.table_name:
             raise NotImplementedError("Sampling of complex queries is not supported")
 
         sample_table = self.unquote_identifier(self.table_name) + '_%s_%s' % (method, str(parameter).replace('.', '_'))
+        create = False
         try:
             with self._execute_sql_query("SELECT * FROM %s LIMIT 0" % self.quote_identifier(sample_table)) as cur:
                 cur.fetchall()
-            with self._execute_sql_query("DROP TABLE %s" % self.quote_identifier(sample_table)) as cur:
-                cur.fetchall()
+
+            if no_cache:
+                with self._execute_sql_query("DROP TABLE %s" % self.quote_identifier(sample_table)) as cur:
+                    cur.fetchall()
+                create = True
+
         except psycopg2.ProgrammingError:
-            pass
-        with self._execute_sql_query('SELECT %s(%s, %s, %s)' % (
-                method,
-                self.quote_string(sample_table),
-                self.quote_string(self.unquote_identifier(self.table_name)),
-                parameter)) as cur:
-            cur.fetchall()
+            create = True
+
+        if create:
+            with self._execute_sql_query('SELECT %s(%s, %s, %s)' % (
+                    method,
+                    self.quote_string(sample_table),
+                    self.quote_string(self.unquote_identifier(self.table_name)),
+                    parameter)) as cur:
+                cur.fetchall()
 
         sampled_table = self.copy()
         sampled_table.table_name = self.quote_identifier(sample_table)
